+++ src/archive/tests/tests.rs
@@ -2,8 +2,8 @@ use canister_tests::api::archive as api;
 use canister_tests::framework::*;
 use ic_test_state_machine_client::CallError;
 use ic_test_state_machine_client::ErrorCode::CanisterCalledTrap;
-use internet_identity_interface::archive::*;
-use internet_identity_interface::*;
+use internet_identity_interface::archive::types::*;
+use internet_identity_interface::internet_identity::types::*;
 use regex::Regex;
 use serde_bytes::ByteBuf;
 use std::time::{Duration, SystemTime};
