+++ src/canister_tests/src/api/archive.rs
@@ -3,8 +3,8 @@ use ic_cdk::api::management_canister::main::CanisterId;
 use ic_test_state_machine_client::{
     call_candid, call_candid_as, query_candid, CallError, StateMachine,
 };
-use internet_identity_interface::archive::*;
-use internet_identity_interface::*;
+use internet_identity_interface::archive::types::*;
+use internet_identity_interface::internet_identity::types::*;
 
 pub fn add_entry(
     env: &StateMachine,
