+++ src/canister_tests/src/framework.rs
@@ -11,9 +11,9 @@ use ic_test_state_machine_client::{CallError, ErrorCode, StateMachine};
 use ic_types::crypto::Signable;
 use ic_types::messages::Delegation;
 use ic_types::Time;
-use internet_identity_interface::archive::*;
+use internet_identity_interface::archive::types::*;
 use internet_identity_interface::http_gateway::{HeaderField, HttpRequest};
-use internet_identity_interface::*;
+use internet_identity_interface::internet_identity::types::*;
 use lazy_static::lazy_static;
 use regex::Regex;
 use serde_bytes::ByteBuf;
