+++ src/internet_identity/src/active_anchor_stats/stats_maintenance.rs
@@ -1,7 +1,7 @@
 use crate::active_anchor_stats::new_active_anchor_counter;
 use crate::DAY_NS;
 use ic_cdk::api::time;
-use internet_identity_interface::ActiveAnchorStatistics;
+use internet_identity_interface::internet_identity::types::ActiveAnchorStatistics;
 
 /// Updates the active anchor counters if an ongoing collection bucket has completed.
 pub fn process_stats(stats: &mut ActiveAnchorStatistics) {
