+++ src/internet_identity/src/archive.rs
@@ -9,8 +9,8 @@ use ic_cdk::api::management_canister::main::{
 };
 use ic_cdk::api::time;
 use ic_cdk::{call, caller, id, trap};
-use internet_identity_interface::archive::*;
-use internet_identity_interface::*;
+use internet_identity_interface::archive::types::*;
+use internet_identity_interface::internet_identity::types::*;
 use serde_bytes::ByteBuf;
 use sha2::Digest;
 use sha2::Sha256;
