+++ src/internet_identity_interface/src/archive/conversions.rs
@@ -1,0 +1,15 @@
+use crate::archive::types::DeviceDataWithoutAlias;
+use crate::internet_identity::types::DeviceData;
+
+impl From<DeviceData> for DeviceDataWithoutAlias {
+    fn from(device_data: DeviceData) -> Self {
+        Self {
+            pubkey: device_data.pubkey,
+            credential_id: device_data.credential_id,
+            purpose: device_data.purpose,
+            key_type: device_data.key_type,
+            protection: device_data.protection,
+            origin: device_data.origin,
+        }
+    }
+}
