+++ src/internet_identity_interface/src/internet_identity.rs
@@ -1,4 +1,4 @@
-/// II archive types compatible on the candid interface.
+/// Internet Identity types compatible on the candid interface.
 /// The compatibility is maintained manually.
 /// Once the support for generating rust types from candid is improved (see https://github.com/dfinity/candid/issues/255)
 /// this module will be generated from the candid interface.
